Fix issues with server cert access and generation
The generation scripts (which I could have avoided using, for all
intents and purposes, maybe by outsourcing to something that handles
certificate generation) change permissions to 0600 after generation the
certificates. To accomodate this, I've changed the scripts to be
executed directly as the user who runs the service.

During the course of that, some other bugs popped up:

* `pki_vars` will only get merged if not overriden from the outside,
  thus:
  * Make all parts of `pki_vars` mandatory (as they are).
  * Use all arguments in example manifest, since the testing environment
    doesn't use Puppet environments with hiera.
* The lookup options weren't defined, which they really should be.